fix(quality): extract trend utils into quality-trend.ts (ESLint fix)
CI lint падал на react-refresh/only-export-components — компонент-файл
не может экспортировать функции, только компоненты. Перенёс утилиты
rolling-avg (FocusMode, bucketPct, lineColor, badgeLabel,
computeRollingAvg) в отдельный pure-data модуль `quality-trend.ts`.

QualityChart реэкспортирует FocusMode для обратной совместимости
импортов (другие callsite'ы могут ссылаться на старый путь).
QualitySummaryPanel импортирует утилиты напрямую из quality-trend.
